Alterações na View e PieChart
